LayoutTests/editing/execCommand/clipboard-access-with-user-gesture-expected.txt

This test checks that JavaScript programs can execute copy and paste commands, but only during a user gesture.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Testing copying an editableParagraph range without a user gesture.
PASS document.queryCommandEnabled('copy') is false
PASS eventSeen['copy'] is false
PASS document.queryCommandSupported('copy') is true
PASS eventSeen['copy'] is false
PASS document.execCommand('copy') is false
PASS eventSeen['copy'] is false

Testing copying an editableParagraph caret without a user gesture.
PASS document.queryCommandEnabled('copy') is false
PASS eventSeen['copy'] is false
PASS document.queryCommandSupported('copy') is true
PASS eventSeen['copy'] is false
PASS document.execCommand('copy') is false
PASS eventSeen['copy'] is false

Testing copying an non-editable range without a user gesture.
PASS document.queryCommandEnabled('copy') is false
PASS eventSeen['copy'] is false
PASS document.queryCommandSupported('copy') is true
PASS eventSeen['copy'] is false
PASS document.execCommand('copy') is false
PASS eventSeen['copy'] is false

Testing copying an non-editable caret without a user gesture.
PASS document.queryCommandEnabled('copy') is false
PASS eventSeen['copy'] is false
PASS document.queryCommandSupported('copy') is true
PASS eventSeen['copy'] is false
PASS document.execCommand('copy') is false
PASS eventSeen['copy'] is false

Testing copying an editable plaint-text range without a user gesture.
PASS document.queryCommandEnabled('copy') is false
PASS eventSeen['copy'] is false
PASS document.queryCommandSupported('copy') is true
PASS eventSeen['copy'] is false
PASS document.execCommand('copy') is false
PASS eventSeen['copy'] is false

Testing copying an editable plaint-text caret without a user gesture.
PASS document.queryCommandEnabled('copy') is false
PASS eventSeen['copy'] is false
PASS document.queryCommandSupported('copy') is true
PASS eventSeen['copy'] is false
PASS document.execCommand('copy') is false
PASS eventSeen['copy'] is false

Testing copying when there is no selection without a user gesture.
PASS document.queryCommandEnabled('copy') is false
PASS eventSeen['copy'] is false
PASS document.queryCommandSupported('copy') is true
PASS eventSeen['copy'] is false
PASS document.execCommand('copy') is false
PASS eventSeen['copy'] is false

Testing cutting an editableParagraph range without a user gesture.
PASS document.queryCommandEnabled('cut') is false
PASS eventSeen['cut'] is false
PASS document.queryCommandSupported('cut') is true
PASS eventSeen['cut'] is false
PASS document.execCommand('cut') is false
PASS eventSeen['cut'] is false

Testing cutting an editableParagraph caret without a user gesture.
PASS document.queryCommandEnabled('cut') is false
PASS eventSeen['cut'] is false
PASS document.queryCommandSupported('cut') is true
PASS eventSeen['cut'] is false
PASS document.execCommand('cut') is false
PASS eventSeen['cut'] is false

Testing cutting an non-editable range without a user gesture.
PASS document.queryCommandEnabled('cut') is false
PASS eventSeen['cut'] is false
PASS document.queryCommandSupported('cut') is true
PASS eventSeen['cut'] is false
PASS document.execCommand('cut') is false
PASS eventSeen['cut'] is false

Testing cutting an non-editable caret without a user gesture.
PASS document.queryCommandEnabled('cut') is false
PASS eventSeen['cut'] is false
PASS document.queryCommandSupported('cut') is true
PASS eventSeen['cut'] is false
PASS document.execCommand('cut') is false
PASS eventSeen['cut'] is false

Testing cutting an editable plaint-text range without a user gesture.
PASS document.queryCommandEnabled('cut') is false
PASS eventSeen['cut'] is false
PASS document.queryCommandSupported('cut') is true
PASS eventSeen['cut'] is false
PASS document.execCommand('cut') is false
PASS eventSeen['cut'] is false

Testing cutting an editable plaint-text caret without a user gesture.
PASS document.queryCommandEnabled('cut') is false
PASS eventSeen['cut'] is false
PASS document.queryCommandSupported('cut') is true
PASS eventSeen['cut'] is false
PASS document.execCommand('cut') is false
PASS eventSeen['cut'] is false

Testing cutting when there is no selection without a user gesture.
PASS document.queryCommandEnabled('cut') is false
PASS eventSeen['cut'] is false
PASS document.queryCommandSupported('cut') is true
PASS eventSeen['cut'] is false
PASS document.execCommand('cut') is false
PASS eventSeen['cut'] is false
PASS successfullyParsed is true

TEST COMPLETE
